+++ enums/EnemyState.h
@@ -1,0 +1,5 @@
+enum class EnemyState
+{
+    ALIVE,
+    DEAD
+};
