+++ enums/GameState.h
@@ -1,0 +1,7 @@
+/// @brief Represents the current game's state
+enum class GameState
+{
+    MENU,
+    PLAY,
+    EXIT
+};
