+++ enums/MenuActions.h
@@ -1,0 +1,7 @@
+enum class MenuActions
+{
+    NONE,
+    START,
+    ABOUT,
+    EXIT
+};
