+++ src/Quiver.cpp
@@ -1,0 +1,16 @@
+#include "../interfaces/Quiver.hpp"
+
+Quiver::Quiver() : Item(QUIVER_IMAGE)
+{
+    arrows = 5;
+}
+
+Quiver::Quiver(int arrows) : Item(QUIVER_IMAGE)
+{
+    this->arrows = arrows;
+}
+
+int Quiver::getArrows()
+{
+    return this->arrows;
+}
