--- src/common.h
@@ -1,14 +1,0 @@
-#ifndef COMMON_H
-#define COMMON_H
-#include <string>
-
-const int GAME_WINDOW_HEIGHT = 800;
-const int GAME_WINDOW_WIDTH = 1200;
-const int PROJECTILE_VELOCITY = 150;
-const std::string GAME_FONT = "resources/fonts/AGoblinAppears-o2aV.ttf";
-const std::string MENU_IMAGE = "resources/assets/bg.png";
-const std::string BACKGROUND_GAME = "resources/assets/background4.png";
-const std::string BASE_IMAGE = "resources/assets/sprites/base.png";
-const std::string HERO_IMAGE = "resources/assets/sprites/hero.png";
-
-#endif
