--- src/enums/GameState.h
@@ -1,7 +1,0 @@
-/// @brief Represents the current game's state
-enum class GameState
-{
-    MENU,
-    PLAY,
-    EXIT
-};
