--- src/enums/MenuActions.h
@@ -1,7 +1,0 @@
-enum class MenuActions
-{
-    NONE,
-    START,
-    ABOUT,
-    EXIT
-};
