+++ src/main.cpp
@@ -1,7 +1,8 @@
-#include "domain/interfaces/Game.hpp"
-#include "domain/interfaces/Menu.hpp"
-#include "enums/GameState.h"
-#include "common.h"
+#include "../interfaces/Game.hpp"
+#include "../interfaces/Menu.hpp"
+#include "../interfaces/Base.hpp"
+#include "../enums/GameState.h"
+#include "../common.h"
 #include "modules/texture_manager/src/ResourceManager.hpp"
 
 const float CENTER_X = GAME_WINDOW_WIDTH / 2;
@@ -16,7 +17,7 @@ int main()
     }
 
     auto newHero = std::make_shared<Hero>(50, 50, 15, 100, 600, 400);
-    auto newBase = std::make_shared<Base>(50, 500, 500, CENTER_X, CENTER_Y);
+    auto newBase = std::make_shared<Base>(500, CENTER_X, CENTER_Y);
     auto window = std::make_shared<sf::RenderWindow>(sf::VideoMode(1200, 800), "Game Window");
     Menu mainMenu(window);
     Game game(CENTER_X, CENTER_Y, window);
